+++ lib/any_cache/error.rb
@@ -11,4 +11,10 @@ class AnyCache
 
   # @since 0.3.1
   NonexistentPatchError = Class.new(Error)
+
+  # @since 0.5.0
+  AlreadyRegisteredPluginError = Class.new(StandardError)
+
+  # @since 0.5.0
+  UnregisteredPluginError = Class.new(StandardError)
 end